Unit test added to Score Applicants page
This unit test tests if the correct number of search results are returned when using a search string. A collection of dummy applicants named John Doe, Jane Doe, and Bob Deer is created. The search string is then set to "doe". 2 search results are expected. If the length of displayedApplicants is 2, the test passes. The search string is then set to "bob". 1 search result is expected. If the length of displayedApplicants is 1, the test passes again. The unit test is then concluded.